split-utils * Move useLocationPermission to hooks/useLocationPermission.ts
