+++ src/containers/LandingPage/SearchPanel/SearchPanel.tsx
@@ -3,9 +3,9 @@ import { Button } from '@entur/button'
 import { Coordinates } from '@entur/sdk'
 import { Dropdown } from '@entur/dropdown'
 import { PositionIcon } from '@entur/icons'
-import { useLocationPermission } from '../../../hooks'
 import { fetchAutocomplete } from '../../../logic/geocoder/fetchAutocomplete'
 import { fetchReverse } from '../../../logic/geocoder/fetchReverse'
+import { useLocationPermission } from '../../../hooks/useLocationPermission'
 import './SearchPanel.scss'
 
 const YOUR_POSITION = 'Posisjonen din'
