+++ src/hooks/useLocationPermission.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react'
+import { useEffect, useState } from 'react'
 
 interface LocationPermission {
     granted: boolean
